Added implementation for the channels.join method
Slack offers the ability to join a room channel in a team as long as the
user is authorized to do so, otherwise an error "restricted_action" will
be returned. If the operation suceeds the response will contain the
status and the information associated with the channel. Notice that if
the channel doesn't exists and the user is authorized to create it the
HTTP request will do so and return the same information on success.