Added implementation for the channels.invite method
Method to allow the invitation of one user at a time to certain channel
as long as the user is authorized to invite with scope "channels:write"
and the invitee exists and is not deleted. The response contains the
result of the action and the information of the channel on success.