Added implementation for the groups.invite method
Method to allow the invitation of one user at a time to certain group
as long as the user is authorized to invite with scope "groups:write"
and the invitee exists and is not deleted. The response contains the
result of the action and the information of the group on success.