Added implementation for the groups.kick method
Method allow to remove one user at a time from a group as long as the
user exists and has joined the group, otherwise an unauthorized error
message is returned in the response. Notice that kicking the same user
that is sending the request is not possible, another user needs to
execute the operation with a different token.