Added implementation for the auth.revoke method
This method revokes an access token. Use it when you no longer need a
token. For example, with a Sign In With Slack app, call this to log a
user out. Notice that "auth.revokeTest" is also available to offer a
method that will simply test the revocation of the token but will keep
the authorization alive.