Added implementation for the chat.meMessage method
This method sends a me message to a channel from the calling user. This
table lists the expected errors that this method could return. However,
other errors can be returned in the case where the service is down or
other unexpected factors affect processing. Callers should always check
the value of the ok params in the response.