Added implementation for the bots.info method
Method to return information about a bot user. The bot identifier is
returned from "bot_message" message events and in the response of
methods like channel history. Notice that a bot user is not the same as
an integration, you will have to find the correct identifier to query
the bot information.