Added implementation for the users.profile.get method
This method is used to get the profile information for a user. If you
call users.profile.get frequently on behalf of a team or user, we
recommend caching labels retrieved from team.profile.get or from when
you sparingly use the include_labels parameter with users.profile.get.
The include_labels parameter is heavily rate-limited. The response
contains a profile item with an array of key:value pairs. The
first_name, last_name, email and skype keys are self-explanatory. The
image_ keys hold links to the different sizes we support for the user's
profile image from 24x24 to 1024x1024 pixels. A link to the image in its
original size is stored in image_original. For a description of the
fields key, see the users.profile.set method.