Added implementation for the team.billableInfo method
This method lists billable information for each user on the team.
Currently this consists solely of whether the user is subject to billing
per Slack's Fair Billing policy. The response contains a list of
activity logs followed by pagination information. A billing_active
status of true indicates that the user is eligible per billing per
Slack's Fair Billing policy. The billing_active status is computed
periodically and the values returned by this API reflect the most
recently computed status and should not be interpreted as a real-time
view of each user's billing_active status.