Added implementation for the users.profile.set method
This method is used to set the profile information for a user.
Individual fields can be updated by passing the pair of arguments name
and value; or multiple fields can be updated at once by passing the
argument profile. The profile argument accepts a URL-encoded JSON object
containing all the fields to be updated. The recommended way to set the
profile image fields is to call users.setPhoto. To clear them, call
users.deletePhoto. The fields key is an array of key:value pairs holding
the values for the user's custom profile fields. The key is the ID of
the definition for that field, which is per-team. This ID can also be
used in the name field of this method. The value of a field is what
should be displayed, unless the alt key is also present, in which case
that is displayed instead. The value can be up to 256 characters for
fields of type text and link. Use team.profile.get to retrieve the
profile fields used by a team.