Added new methods to set the user avatar and omit errors
Some of the Slack API endpoints are hidden in the documentation, this
includes the methods used to upload a picture and set it as the avatar,
as well as the method to delete it. This commit adds these placeholders
to be later implemented, and adds additional tags for the default
response object to omit empty error messages.