Added implementation for the chat.robotMessage method
Allows to send messages to a specific channel as a robot. You will have
to generate a private API key (aka. xoxp Slack token) as the common key
ignores the "as_user" parameter which is the one used by the Slack API
service to determine if the message will be displayed as sent by a robot
and also have to set two additional environment variables named
SLACK_ROBOT_NAME and SLACK_ROBOT_IMAGE which can be an URL pointing to a
JPEG, GIF, or PNG file or the name of a regular emoji prefixed with a
colon character.